convert eviction set to use shrink set

+++ src/utils/shrink.rs
@@ -1,9 +1,10 @@
 use std::{
   collections::{vec_deque::Drain, VecDeque},
+  hash::{BuildHasher, Hash},
   ops::{Deref, DerefMut, RangeBounds},
 };
 
-use hashbrown::raw::RawTable;
+use hashbrown::{raw::RawTable, Equivalent};
 
 pub struct HashTable<T>(RawTable<T>);
 impl<T> HashTable<T> {
@@ -17,6 +18,9 @@ impl<T> HashTable<T> {
     eq: impl FnMut(&T) -> bool,
     hasher: impl Fn(&T) -> u64,
   ) -> Option<T> {
+    if self.0.is_empty() {
+      return None;
+    }
     let v = self.0.remove_entry(hash, eq)?;
 
     let cap = self.0.capacity();
@@ -38,6 +42,36 @@ impl<T> Deref for HashTable<T> {
 impl<T> DerefMut for HashTable<T> {
   fn deref_mut(&mut self) -> &mut Self::Target {
     &mut self.0
+  }
+}
+
+pub struct HashSet<K>(HashTable<K>);
+impl<K> HashSet<K> {
+  pub const fn new() -> Self {
+    Self(HashTable::new())
+  }
+  pub fn contains<Q>(&self, hash: u64, key: &Q) -> bool
+  where
+    Q: Equivalent<K> + ?Sized,
+  {
+    self.0.find(hash, |k| key.equivalent(k)).is_some()
+  }
+  pub fn insert<S>(&mut self, key: K, hash: u64, hasher: &S)
+  where
+    K: Hash,
+    S: BuildHasher,
+  {
+    self.0.insert(hash, key, |k| hasher.hash_one(k));
+  }
+  pub fn remove<Q, S>(&mut self, hash: u64, key: &Q, hasher: &S)
+  where
+    Q: Equivalent<K> + ?Sized,
+    K: Hash,
+    S: BuildHasher,
+  {
+    self
+      .0
+      .remove_and_shrink(hash, |k| key.equivalent(k), |k| hasher.hash_one(k));
   }
 }
